SEQNG-654 Fixed PR comments
