+++ modules/seqexec/server/src/main/scala/seqexec/server/tcs/Tcs.scala
@@ -13,7 +13,6 @@ import edu.gemini.spModel.seqcomp.SeqConfigNames.TELESCOPE_KEY
 import edu.gemini.spModel.target.obsComp.TargetObsCompConstants._
 import org.log4s.getLogger
 import mouse.all._
-import monocle.Prism
 import scala.language.implicitConversions
 import scala.reflect.ClassTag
 import seqexec.model.Model.Resource
@@ -103,28 +102,6 @@ object Tcs {
   val GUIDE_WITH_AOWFS_PROP: String = "guideWithAOWFS"
   val P_OFFSET_PROP: String = "p"
   val Q_OFFSET_PROP: String = "q"
-
-  sealed trait CRFollow
-  object CRFollow {
-    case object On extends CRFollow
-    case object Off extends CRFollow
-
-    def keywordValue(cr: CRFollow): String = cr match {
-      case On  => "yes"
-      case Off => "no"
-    }
-
-    def fromInt: Prism[Int, CRFollow] =
-      Prism[Int, CRFollow] {
-        case 0  => Off.some
-        case 1  => On.some
-        case _  => none
-      } {
-        case Off => 0
-        case On  => 1
-      }
-
-  }
 
   // Conversions from ODB model values to TCS configuration values
   implicit def probeTrackingConfigFromGuideWith(guideWith: StandardGuideOptions.Value): ProbeTrackingConfig = guideWith match {
