+++ modules/seqexec/server/src/test/scala/seqexec/server/tcs/TcsSpec.scala
@@ -6,6 +6,7 @@ package seqexec.server.tcs
 import cats.kernel.laws.discipline._
 import cats.tests.CatsSuite
 import edu.gemini.seqexec.server.tcs.{BinaryOnOff, BinaryYesNo}
+import monocle.law.discipline.PrismTests
 
 /**
   * Tests Tcs typeclasses
@@ -15,4 +16,6 @@ final class TcsSpec extends CatsSuite {
 
   checkAll("Eq[BinaryYesNo]", EqTests[BinaryYesNo].eqv)
   checkAll("Eq[BinaryOnOff]", EqTests[BinaryOnOff].eqv)
+  checkAll("Eq[CRFollow]", EqTests[CRFollow].eqv)
+  checkAll("Prism[Int, CRFollow]", PrismTests(CRFollow.fromInt))
 }
